+++ e2e/actions/actions.go
@@ -2635,5 +2635,6 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"ociOverlay":           (c.actionOciOverlay),           // --overlay in OCI mode
 		"ociOverlayExtfsPerms": (c.actionOciOverlayExtfsPerms), // permissions in writable extfs overlays mounted with FUSE in OCI mode
 		"ociOverlayTeardown":   np(c.actionOciOverlayTeardown), // proper overlay unmounting in OCI mode
+		"ociNo-mount":          c.actionOciNoMount,             // --no-mount in OCI mode
 	}
 }